% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_communities-data.R
\docType{data}
\name{four_communities-data}
\alias{four_communities-data}
\alias{four_communities}
\title{Vegetation-Plots Sampled in Four Syntaxa}
\format{
An object of class \code{list} of length 4.
}
\usage{
four_communities
}
\description{
Data set collected in several stands of vegetation belonging to four
phytosociological, namely Eleocharito-Caricetum bohemicae (class
Isoëto-Nanojuncetea), Gentiano-Koelerietum pyramidetae (class
Festuco-Brometea), Mesobrometum (class Festuco-Brometea), and
Alliario-Chaerophylletum temuli (class Galio-Urticetea).

The data set is stored as a list, including four data frames: \code{Head}
whit plots information, \code{RelevesList} including the abundance of
species in single plots (as a column oriented table), \code{SpeciesList}
with the information of single species, and \code{CoverConversion} with the
conversion table for Braun-Blanquet abundance to percentage cover.

This data set is offered as alternative for examples working with
Ellenberg's indicator values (EIV) and life forms. The analysis of those
data was published in (\bold{Alvarez et al. 2013}).
}
\references{
\bold{Alvarez M, Möseler BM, San Martín C (2013).} Tendency
and distribution indices for ordinal variables applied to Ellenberg
indicator values.  \emph{Gayana Botánica} 70: 225--233.
}
\keyword{datasets}
