% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexiguel-package.R
\docType{package}
\name{Lexiguel-package}
\alias{Lexiguel-package}
\alias{Taxlist}
\title{A Mingle-Mangle of Functions for R.}
\description{
Most of the functions included here are related with published works and
on-going projects carried out by the maintainer. This package is currently
hosted by GitHub (\url{https://github.com/kamapu/Lexiguel}). For its
installation, you may first install the package \code{devtools} and dependencies,
and then follow the provided example.
}
\examples{
\dontrun{
library(devtools)
install_github("kamapu/Lexiguel")
}

}
\author{
Miguel Alvarez (\email{malvarez@uni-bonn.de}).
}
