% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lim_bars.R
\name{lim_bars}
\alias{lim_bars}
\title{Limits bar plot}
\usage{
lim_bars(
  x,
  y,
  breaks = 10,
  pos = NULL,
  horiz = FALSE,
  poly_args = list(),
  pos_args = list(),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{x}{Either a numeric vector or a matrix with two columns.}

\item{y}{A numeric vector. If missing the second column of 'x' will be
used instead.}

\item{breaks}{An integer value indicating the number of segments used to
split 'x'.}

\item{pos}{A function used to determine the position of the segment. For
instance 'mean' or 'median' can be used. By default no position will be
displayed.}

\item{horiz}{A logical value indicating whether bars should be drawn
horizontally or vertically.}

\item{poly_args}{A list including arguments to be passed to \code{\link[=polygon]{polygon()}}. This
will affect the layout of the bars.}

\item{pos_args}{A list including arguments to be passed to \code{\link[=segments]{segments()}}. This
will affect the layout of the position lines.}

\item{na.rm}{A logical value indicating whether NA values in 'x' and 'y'
should removed or not.}

\item{...}{Further arguments passed to \code{\link[=plot]{plot()}}.}
}
\description{
Graphic of bars showing the range (minimum and maximum values) of a variable
'y' along segments of a second 'x' variable.
}
