% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_deps.R
\name{recursive_deps}
\alias{recursive_deps}
\title{List Package Dependencies Recursively}
\usage{
recursive_deps(packages, ...)
}
\arguments{
\item{packages}{A character vector with names of packages.}

\item{...}{Further arguments passed to \code{\link[=package_dependencies]{package_dependencies()}}}
}
\value{
A character vector with the name of all dependencies required recursively.
}
\description{
In order to recognize all requirements that needs to be installed in a fresh
system, you need to collect information about dependencies recursively.
This package is a wrapper for \code{\link[=package_dependencies]{package_dependencies()}}.

The packages do not necessarily need to be installed locally, they only need
to be available from CRAN.
}
