% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowplot.R
\name{shadowplot}
\alias{shadowplot}
\alias{shadows}
\alias{shadowplot.default}
\alias{shadowplot.numeric}
\alias{shadowplot.data.frame}
\alias{shadowplot.formula}
\title{Shadowplots for categorical variables}
\usage{
shadows(x, start, end, at, maxwidth = 1, col = "grey", border = NA, ...)

shadowplot(x, ...)

\method{shadowplot}{default}(x, ...)

\method{shadowplot}{numeric}(
  x,
  start,
  end,
  at,
  maxwidth = 1,
  col = "grey",
  border = NA,
  bty = "n",
  xaxt = "n",
  ylim,
  xlab = "",
  ylab = "classes",
  ...
)

\method{shadowplot}{data.frame}(
  x = data.frame(),
  what,
  by,
  start,
  end,
  at,
  maxwidth = 1,
  col = "grey",
  border = NA,
  xlab = "",
  ylab = "classes",
  xlim,
  ylim,
  xaxt = "s",
  ...
)

\method{shadowplot}{formula}(formula, data = data.frame(), ...)
}
\arguments{
\item{x}{Either an integer vector or a data frame.}

\item{start}{Lowest possible value for the variable 'x' as integer.}

\item{end}{Highest possible value for the variable 'x' as integer.}

\item{at}{Position of the shadow in the x-axis of the plot.}

\item{maxwidth}{Magnification factor for the maximum width of shadow.}

\item{col}{Color of the shadow (passed to \code{\link[=polygon]{polygon()}}).}

\item{border}{Color of the shadow's border (passed to \code{\link[=polygon]{polygon()}}).}

\item{...}{Further arguments to be passed among methods.}

\item{bty}{Type of box around plotting area (see \code{\link[=par]{par()}}).}

\item{xaxt}{Specification of x axis type (see \code{\link[=par]{par()}}).}

\item{xlab}{Labels for x axis (see \code{\link[=par]{par()}}).}

\item{ylab}{Labels for y axis (see \code{\link[=par]{par()}}).}

\item{what}{Name of the variable to be plotted (in data frame method).}

\item{by}{Name of the splitting variable (in data frame method).}

\item{xlim, ylim}{Limits for x and y axes (see \code{\link[=par]{par()}}).}

\item{formula}{A formula for plotting.}

\item{data}{Data frame containing the vectors to be plotted.}
}
\description{
Plot single 'shadow'. This function is called by \code{shadowplot()} and works
only as a low-level plot function.

The function \code{shadow()} calculates the frequency of single values in the
vector \code{x} and displays them as a shadow using the function \code{\link[=polygon]{polygon()}}.
The resulting shadow indicates by its with the relative frequency of the
respective values in the input vector.
If not provided, the values of \code{start} and \code{end} will be calculated by
\code{\link[=min]{min()}} and \code{\link[=max]{max()}}, respectively.
This plot function is offered as an alternative to \code{\link[=boxplot]{boxplot()}} for
categorical and ordinal data and should be interpreted as an histogram.

The function \code{shadowplot()} produces a plot of frequencies for classes of
categorical variables and is similar to \emph{box-and-whisker plots} \code{\link[=boxplot]{boxplot()}})
but also to \emph{histograms} \code{\link[=hist]{hist()}}).

This plot function is offered as an alternative to \code{\link[=boxplot]{boxplot()}} for
categorical and ordinal data and single shadows should be interpreted as
histograms. To plot single shadows the function \code{shadows()} is called
internally.
Alternatively \code{\link[=stringplot]{stringplot()}} or a combination of both can be used.
}
\examples{
## Similar example as for strp
data(WHscores)

## Shadow overlaid to box and whisker plot using transparent color
## (note the bimodal distribution)
boxplot(WHscores$WET_veg, col="grey")
shadows(WHscores$WET_veg, col=rgb(153/255, 255/255, 153/255, alpha=0.5),
    at=1)

## Plot using formula
shadowplot(WET_veg ~ Landuse, WHscores, start=0, end=10,
        col=c("orangered", "palegreen", "orange", "lightblue"))

## Adding a stringplot
stringplot(WET_veg ~ Landuse, WHscores, lwd=3, at=1)

}
\author{
Miguel Alvarez (\email{malvarez@uni-bonn.de}).
}
