% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_groups.R
\name{signif_groups}
\alias{signif_groups}
\title{Retrieve significant groups after pairwise post-hoc tests}
\usage{
signif_groups(g, m, t_tab, criterion, decreasing = TRUE)
}
\arguments{
\item{g}{A character vector with the names of the groups.}

\item{m}{A numeric vector with the average values for the tested variable in
each group. This vector is used to sort the groups (from bigger to smaller
values) and assign the respective letters.}

\item{t_tab}{A data frame with the results of pairwise contrast. Two columns
are mandatory in this table, namely 'group1' and 'group2' containing the
contrasted groups. Note that the values in these columns need to match 'g'.}

\item{criterion}{An expression (symbol) used to define groups belonging to
the same significant group, for instance a cut level for a p-value.}

\item{decreasing}{Logical value, whether the sorting of mean values should be
done from the highest to the lowest or vice versa. This argument is passed
to \code{\link[=order]{order()}}.}
}
\value{
A data frame with the names of the groups and their respective significant
groups.
}
\description{
Retrieve significant groups after a pairwise post-hoc test.
}
