% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringplot.R
\name{stringplot}
\alias{stringplot}
\alias{strp}
\alias{stringplot.default}
\alias{stringplot.numeric}
\alias{stringplot.data.frame}
\alias{stringplot.formula}
\title{Plot for Distribution of Categorical Variables}
\usage{
strp(x, lwd = 1, maxwidth = 1, col = "black", at)

stringplot(x, ...)

\method{stringplot}{default}(x, ...)

\method{stringplot}{numeric}(
  x,
  lwd = 1,
  maxwidth = 1,
  col = "black",
  at,
  bty = "n",
  xaxt = "n",
  ylim,
  xlab = "",
  ylab = "classes",
  ...
)

\method{stringplot}{data.frame}(
  x = data.frame(),
  what,
  by,
  at,
  col,
  lwd = 2,
  maxwidth = 1,
  xlab = "",
  ylab = "classes",
  xlim,
  ylim,
  xaxt = "s",
  ...
)

\method{stringplot}{formula}(formula, data = data.frame(), ...)
}
\arguments{
\item{x}{Either an integer vector or a data frame.}

\item{lwd}{Line width for strings (see \code{\link[=par]{par()}}).}

\item{maxwidth}{Magnification factor for line width.}

\item{col}{Colour for strings.}

\item{at}{Position of string piles in existing plot.}

\item{...}{Further arguments to be passed among methods.}

\item{bty}{Type of box around plotting area (see \code{\link[=par]{par()}}).}

\item{xaxt}{Specification of x axis type (see \code{\link[=par]{par()}}).}

\item{xlab}{Labels for x axis (see \code{\link[=par]{par()}}).}

\item{ylab}{Labels for y axis (see \code{\link[=par]{par()}}).}

\item{what}{Name of the variable to be plotted (in data frame method).}

\item{by}{Name of the splitting variable (in data frame method).}

\item{xlim, ylim}{Limits for x and y axes (see \code{\link[=par]{par()}}).}

\item{formula}{A formula for plotting.}

\item{data}{Data frame containing the vectors to be plotted.}
}
\description{
Alternative plot to display distribution of categorical variables.
The function \code{stringplot()} produces a plot of frequencies for classes of
categorical variables and is similar to \emph{box-and-whisker plots} but
it may be interpreted as histograms.

This plot function is offered as an alternative to \code{\link[=boxplot]{boxplot()}} for
categorical (ordinal) variables and single string piles should be interpreted
as histograms.
To plot single string piles the function \code{\link[=strp]{strp()}} is called internally.
Alternatively \code{\link[=shadowplot]{shadowplot()}} or a combination of both can be used.

The function \code{strp()} calculates the frequency of single values in the vector
\code{x} and displays them in a string pile by using the function \code{\link[=segments]{segments()}}.
The resulting pile indicates by the with of strings the relative frequency of
the respective values in the input vector.
This plot function is offered as an alternative to \code{\link[=boxplot]{boxplot()}} for
categorical and ordinal data and should be interpreted as an histogram.
}
\examples{
data(WHscores)

## Hydrology scores per land use
stringplot(WET_hydro ~ Landuse, data=WHscores, xlab="Land uses",
	   ylab="WET-health scores", main="Hydrology module", lwd=3)

## Adding tendencies
Tendency <- aggregate(WET_hydro ~ Landuse, WHscores, median)
with(Tendency, {
    lines(1:4, WET_hydro, lty=2)
    points(1:4, WET_hydro, pch=21, bg="white", cex=1.5)
})

#' ## Strings overlaid to box and whisker plot (note the bimodal distribution)
boxplot(WHscores$WET_veg, border="grey", col="palegreen")
strp(WHscores$WET_veg, lwd=4, at=1)

}
\references{
\strong{Beuel S, Alvarez M, Amler E, Behn K, Kotze D, Kreye C, Leemhuis C,
Wagner K, Willy DK, Ziegler S, Becker M (2016).}
A rapid assessment of anthropogenic disturbances in East African wetlands.
\emph{Ecological Indicators} 67: 684-692.
}
\seealso{
\code{\link{strp}}, \code{\link{shadowplot}}.

\code{\link[=stringplot]{stringplot()}}, \code{\link[=error_bars]{error_bars()}}.
}
\author{
Miguel Alvarez (\email{malvarez@uni-bonn.de}).

Miguel Alvarez (\email{malvarez@uni-bonn.de}).
}
