% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab4net.R
\name{tab4net}
\alias{tab4net}
\title{Prepare inputs for sankeyNetwork()}
\usage{
tab4net(formula, data, FUN, ...)
}
\arguments{
\item{formula}{A formula passed to
\code{\link[stats:aggregate]{stats::aggregate()}}.}

\item{data}{A data frame with the variables included in 'formula'.}

\item{FUN}{A function passed to
\code{\link[stats:aggregate]{stats::aggregate()}}.}

\item{...}{Further arguments passed to
\code{\link[stats:aggregate]{stats::aggregate()}}.}
}
\value{
A list with two data frames called 'nodes' and 'links'.
}
\description{
This function generates lists from column-oriented tables (data frames) for
Sankey Networks diagrams produced by
\code{\link[networkD3:sankeyNetwork]{networkD3::sankeyNetwork()}}.
}
\examples{

require(networkD3)

## Different values of WET-Health scores for vegetation per land use
veg_wh <- tab4net(UseunitID ~ Landuse + WET_veg, WHscores, length)

sankeyNetwork(Links=veg_wh$links, Nodes=veg_wh$nodes, Source = 'source',
    Target = 'target', Value = 'value', NodeID = 'name',
    units = 'TWh', fontSize = 12, nodeWidth = 30)

}
