\name{four_communities-data}
\docType{data}
\alias{four_communities}
\alias{four_communities-data}

\title{Vegetation-Plots Sampled in Four Syntaxa}
\description{
Data set collected in several stands of vegetation belonging to four
phytosociological, namely Eleocharito-Caricetum bohemicae (class
Isoëto-Nanojuncetea), Gentiano-Koelerietum pyramidetae (class Festuco-Brometea),
Mesobrometum (class Festuco-Brometea), and Alliario-Chaerophylletum temuli
(class Galio-Urticetea).
}
\usage{
data(four_communities)
}
\details{
The data set is stored as a list, including four data frames: \code{Head} whit
plots information, \code{RelevesList} including the abundance of species in
single plots (as a column oriented table), \code{SpeciesList} with the
information of single species, and \code{CoverConversion} with the conversion
table for Braun-Blanquet abundance to percentage cover.

This data set is offered as alternative for examples working with Ellenberg's
indicator values (EIV) and life forms.
The analysis of those data was published in (\bold{Alvarez et al. 2013}).
}
\references{
\bold{Alvarez M, Möseler BM, San Martín C (2013).} Tendency and distribution
indices for ordinal variables applied to Ellenberg indicator values. 
\emph{Gayana Botánica} 70: 225--233.
}
