% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_gaps-data.R
\docType{data}
\name{rc_gaps-data}
\alias{rc_gaps-data}
\alias{rc_gaps}
\alias{rc_gaps.env}
\alias{rc_gaps.env-data}
\title{Plant Diversity in Forest Gaps on the Robinson Crusoe Island}
\description{
Data set with floristic composition, size and type of gaps sampled in forest
stands of the Robinson Crusoe Island (Chile).

The data is contained in two separated objects.
The vegetation table (\code{rc_gaps}) is formatted as a presence/absence matrix
with the gaps (units) as rows and the species as columns.
The data frame \code{rc_gaps.env} contains information about the surface size of
the gaps (in squared metres), the type and the size rank (see
\strong{Vargas et al. 2013}).
}
\references{
\strong{Vargas RI, Gärtner S, Alvarez M, Hagen E, Reif A (2013).}
Does restoration help the conservation of the threatened forest of Robinson
Crusoe Island? The impact of forest gap attributes on endemic plant species
richness and exotic invasions.
\emph{Biodiversity and Conservation} 22: 1283-1300.
}
